Polish
Move visitor into recipe, as it's not reused
Add additional tests & apply formatter
Collapse conditionals into single if statement
Remove move imports
